requirements.txt
alabaster==0.7.12
Babel==2.9.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
docutils==0.16
funcsigs==1.0.2
idna==3.2
imagesize==1.2.0
Jinja2==3.0.1
mock==4.0.3
packaging==21.0
Pillow==8.3.1
geocoder
font-source-sans-pro
Pygments==2.9.0
pyparsing==2.4.7
inky
pytz==2021.1
requests==2.26.0
six==1.16.0
snowballstemmer==2.1.0
Sphinx==4.1.2
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib-websupport==1.2.4
typing==3.7.4.3
urllib3==1.26.6
